requirements.txt
python-pptx>=1.0.0
Pillow>=9.0.0
